// frontend/src/api/backend/models/NewVivienda.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tercer Parcial - API REST. Ingeniería Web
 * José Luis Bueno Pachón.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NewVivienda
 */
export interface NewVivienda {
    /**
     * 
     * @type {string}
     * @memberof NewVivienda
     */
    titulo: string;
    /**
     * 
     * @type {string}
     * @memberof NewVivienda
     */
    descripcion: string;
    /**
     * 
     * @type {string}
     * @memberof NewVivienda
     */
    user: string;
    /**
     * 
     * @type {number}
     * @memberof NewVivienda
     */
    precio: number;
    /**
     * 
     * @type {string}
     * @memberof NewVivienda
     */
    direccion: string;
    /**
     * 
     * @type {number}
     * @memberof NewVivienda
     */
    latitude: number;
    /**
     * 
     * @type {number}
     * @memberof NewVivienda
     */
    longitude: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof NewVivienda
     */
    images?: Array<string>;
}

/**
 * Check if a given object implements the NewVivienda interface.
 */
export function instanceOfNewVivienda(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "titulo" in value;
    isInstance = isInstance && "descripcion" in value;
    isInstance = isInstance && "user" in value;
    isInstance = isInstance && "precio" in value;
    isInstance = isInstance && "direccion" in value;
    isInstance = isInstance && "latitude" in value;
    isInstance = isInstance && "longitude" in value;

    return isInstance;
}

export function NewViviendaFromJSON(json: any): NewVivienda {
    return NewViviendaFromJSONTyped(json, false);
}

export function NewViviendaFromJSONTyped(json: any, ignoreDiscriminator: boolean): NewVivienda {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'titulo': json['titulo'],
        'descripcion': json['descripcion'],
        'user': json['user'],
        'precio': json['precio'],
        'direccion': json['direccion'],
        'latitude': json['latitude'],
        'longitude': json['longitude'],
        'images': !exists(json, 'images') ? undefined : json['images'],
    };
}

export function NewViviendaToJSON(value?: NewVivienda | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'titulo': value.titulo,
        'descripcion': value.descripcion,
        'user': value.user,
        'precio': value.precio,
        'direccion': value.direccion,
        'latitude': value.latitude,
        'longitude': value.longitude,
        'images': value.images,
    };
}